Make "publiccode" required and always show the field.
In line with the position paper from the College voor de Rechten van de Mens: Make is visible when transparency about the detailed logic behind the algorithm isn't available. And because intellectual property may not be an obstacle to this transparency I suggest making this field required as well. Under exceptional circumstances this linkt could also link to a resource that's (partially) redacted.